requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.1
pydantic-settings==2.1.0
firebase-admin==6.2.0
httpx==0.25.1
python-dotenv==1.0.0
langdetect==1.0.9
